- mercs only complain about others' diseases after an outbreak - a radio set can function without batteries if the xml states batteries are not required - changed 'friendly fire' dynamic dialogue to not only fit firearm-related incidents
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@7861 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Items.cpp
@@ -8095,7 +8095,7 @@ BOOLEAN CreateItem( UINT16 usItem, INT16 bStatus, OBJECTTYPE * pObj )
 		// Flugente: certain items get the maximum temperature when they are created
 		if ( HasItemFlag(usItem, POWER_PACK) )
 		{
-			(*pObj)[0]->data.bTemperature = (FLOAT)(OVERHEATING_MAX_TEMPERATURE);
+			(*pObj)[0]->data.bTemperature = OVERHEATING_MAX_TEMPERATURE;
 		}
 
 		//ADB ubWeight has been removed, see comments in OBJECTTYPE
